Fix bugs in heroes demo (as part of U3 migration).

+++ src/foam/demos/heroes/Hero.js
@@ -10,7 +10,7 @@ foam.CLASS({
 
   properties: [
     {
-      class: 'Int',
+      class: 'Long',
       name: 'id',
       final: true // TODO: implement
     },
